+++ package/src/decorator/decDebounce.ts
@@ -2,7 +2,7 @@ import { toDecorator } from "@decorator/toDecorator.js";
 import { debounce } from "@function/debounce.js";
 
 /**
- * Debouces the decorated function. Only calling it after a specified amount of time has passed without any new calls.
+ * Debounces the decorated function. Only calling it after a specified amount of time has passed without any new calls.
  * 
  * Look at {@link debounce} for the non-decorator version.
  *